SIDE-2143: added mapper tests for updating upon change
